% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{s_num}
\alias{s_num}
\title{Get Number}
\usage{
s_num(x_name, sub = character(0))
}
\arguments{
\item{x_name}{A string of the name.}

\item{sub}{A string specifying the path to the sub folder (by default the current sub folder).}
}
\value{
A number.
}
\description{
Get Number
}
