% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{s_str}
\alias{s_str}
\title{Get String}
\usage{
s_str(x_name, sub = character(0))
}
\arguments{
\item{x_name}{A string of the name.}

\item{sub}{A string specifying the path to the sub folder (by default the current sub folder).}
}
\value{
A string.
}
\description{
Get String
}
