% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{sbr_figures}
\alias{sbr_figures}
\title{Figures}
\usage{
sbr_figures(sub = character(0), drop = NULL, sort = NULL,
  rename = NULL, nheaders = 2L, header1 = 4L, overwrite = TRUE,
  main = subfoldr2::sbf_get_main(), width = 6)
}
\arguments{
\item{sub}{A string of the path to the root sub folder.}

\item{drop}{A character vector specifying the sub folders and file names to exclude from the report or NULL.}

\item{sort}{A character vector specifying the initial sort order for sub folders and file names or NULL.
Missing items appear afterwards in alphabetical order.}

\item{rename}{A unique named character vector specifying new heading names for sub folders or NULL.
By default sub folder names have the first letter of each word capitalized.}

\item{nheaders}{A count of the number of sub folder levels to assign headers to.}

\item{header1}{A count of the heading level for the first sub folder level.}

\item{overwrite}{A flag specifying whether to overwrite existing files in the report folder.}

\item{main}{A string of the path to the main folder.}

\item{width}{A number of the page width in inches.}
}
\value{
A string of the plots in markdown format.
}
\description{
Returns a string of the report plots and windows in markdown format ready for inclusion in a report.
If a window and plot have identical names then the plot is dropped.
}
