% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{sbr_knit_report}
\alias{sbr_knit_report}
\title{Knit Report}
\usage{
sbr_knit_report(file = "report", report = sbr_get_report(),
  sub = subfoldr2::sbf_get_sub(), main = subfoldr2::sbf_get_main(),
  quiet = FALSE, browse = TRUE)
}
\arguments{
\item{file}{A string of the output file name without an extension.}

\item{report}{A string of the directory to save the output file.}

\item{sub}{A string specifying the path to the sub folder (by default the current sub folder).}

\item{main}{A string specifying the path to the main folder (by default the current main folder)}

\item{quiet}{An option to suppress printing of the pandoc command line.}

\item{browse}{A flag specifying whether to open the html file.}
}
\value{
An invisible path to the .html file.
}
\description{
Knit's tables, code blocks and plots (and windows) into a html file and 
if browse = TRUE opens in the default web browser.
}
