% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{sbr_number}
\alias{sbr_number}
\alias{sbr_n}
\title{Get Number}
\usage{
sbr_number(x_name, sub = character(0))

sbr_n(x_name, sub = character(0))
}
\arguments{
\item{x_name}{A string of the name.}

\item{sub}{A string specifying the path to the sub folder (by default the
current sub folder).}
}
\value{
A number.
}
\description{
Get Number
}
