% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{sbr_set_report}
\alias{sbr_set_report}
\title{Set Report}
\usage{
sbr_set_report(..., rm = FALSE, ask = getOption("sbf.ask", TRUE))
}
\arguments{
\item{...}{One or more character vectors which are combined together.}

\item{rm}{A flag specifying whether to remove the folder and all its contents if it already exists.}

\item{ask}{A flag specifying whether to ask before removing the existing folder.}
}
\value{
An invisible string of the path to the main folder.
}
\description{
The directory is created when needed 
 if it doesn't already exist.
}
