% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{sbr_string}
\alias{sbr_string}
\alias{sbr_s}
\title{Get String}
\usage{
sbr_string(x_name, sub = character(0))

sbr_s(x_name, sub = character(0))
}
\arguments{
\item{x_name}{A string of the name.}

\item{sub}{A string specifying the path to the sub folder (by default the
current sub folder).}
}
\value{
A string.
}
\description{
Get String
}
