% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{sbr_use_knit_report_all}
\alias{sbr_use_knit_report_all}
\title{Use Knit Report All}
\usage{
sbr_use_knit_report_all()
}
\description{
Creates knit-report.R, bibliography.bib and and knit-report.R files
in the root directory.
}
