% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subreport-package.R
\docType{package}
\name{subreport-package}
\alias{subreport}
\alias{subreport-package}
\title{subreport: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
