% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subreport-package.R
\docType{package}
\name{subreport-package}
\alias{subreport}
\alias{subreport-package}
\title{subreport: Generate Reports from `subfoldr2` Objects}
\description{
Facilitates generation of .Rmd and .html reports from tables, (code) blocks and figures (plots and windows) saved using the subfoldr2 package. To quickly view all the results use `sbr_knit_results()`. The `sbr_n()` and `sbr_s()` functions can be used to insert numbers or strings saved using subfoldr2.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
